requirements.txt
# Core AI/ML Libraries
openai>=1.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
neo4j>=5.0.0
langchain>=0.1.0
llama-index>=0.9.0

# MLOps & Pipeline
zenml>=0.50.0
mlflow>=2.8.0

# Web Scraping & Data Ingestion
scrapy>=2.11.0
newspaper3k>=0.2.8

# API & Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Monitoring & Observability
prometheus-client>=0.19.0

# UI & Visualization
streamlit>=1.28.0
d3py>=0.0.2

# Development & Testing
pytest>=7.4.0
flake8>=6.1.0
black>=23.0.0
isort>=5.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0 